Make "artist.address" non-mandatory
As it looks for now, it's probable that we won't be able to match every
single artist's address on the various music NFT platforms. While many
platforms say that only an artist should mint the NFT, that to us
doesn't really indicate that they did it - or e.g. a person/organization
acting as a proxy.

We, hence, think it's more important to correctly identify those
platforms that explicitly log an artists address. For all others, we
think that e.g. the "erc721.minting.from" field is potentially more
appropriate.